Add Test for WaitForReady and WaitForExist.
