+++ tests/IC.Navigation.UnitTests/NavigableExTests/WaitForReady/INavigable_CancellationToken_.cs
@@ -9,7 +9,7 @@ using System;
 using System.Threading;
 using Xunit;
 
-namespace IC.Navigation.UnitTests.NavigableExTests.WaitForExist
+namespace IC.Navigation.UnitTests.NavigableExTests.WaitForReady
 {
     public class INavigable_CancellationToken_
     {
@@ -22,10 +22,10 @@ namespace IC.Navigation.UnitTests.NavigableExTests.WaitForExist
                 var fixture = new Fixture().Customize(new AutoMoqCustomization());
                 using var cancellationTokenSource = new CancellationTokenSource(TimeSpan.FromMilliseconds(1000));
                 var cancellationToken = cancellationTokenSource.Token;
-                Mock.Get(navigable).Setup(x => x.NavigatorSession.WaitForExist(navigable, cancellationToken)).Verifiable();
+                Mock.Get(navigable).Setup(x => x.NavigatorSession.WaitForReady(navigable, cancellationToken)).Verifiable();
 
                 // Act
-                var actual = navigable.WaitForExist(cancellationToken);
+                var actual = navigable.WaitForReady(cancellationToken);
 
                 // Assert
                 actual.Should().BeTrue();
@@ -41,10 +41,10 @@ namespace IC.Navigation.UnitTests.NavigableExTests.WaitForExist
                 var fixture = new Fixture().Customize(new AutoMoqCustomization());
                 using var cancellationTokenSource = new CancellationTokenSource(TimeSpan.Zero);
                 var cancellationToken = cancellationTokenSource.Token;
-                Mock.Get(navigable).Setup(x => x.NavigatorSession.WaitForExist(navigable, cancellationToken)).Verifiable();
+                Mock.Get(navigable).Setup(x => x.NavigatorSession.WaitForReady(navigable, cancellationToken)).Verifiable();
 
                 // Act
-                var actual = navigable.WaitForExist(cancellationToken);
+                var actual = navigable.WaitForReady(cancellationToken);
 
                 // Assert
                 actual.Should().BeFalse();
